backend/pms-core/requirements.txt
blessed==1.20.0
botocore==1.34.43
cachetools==5.3.2
certifi==2024.2.2
chalice==1.30.0
charset-normalizer==3.3.2
click==8.1.7
google-auth==2.27.0
idna==3.6
inquirer==2.10.1
jmespath==1.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0.1
readchar==4.0.5
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.13
